requirements.txt
astroid==2.2.5
certifi==2019.9.11
chardet==3.0.4
colorama==0.4.1
defusedxml==0.6.0
Django==2.2.5
django-allauth==0.40.0
djangorestframework==3.10.3
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.2
mccabe==0.6.1
oauthlib==3.1.0
Pillow==6.1.0
pylint==2.3.1
pylint-django==2.0.11
pylint-plugin-utils==0.5
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.4.0
urllib3==1.25.5
wrapt==1.11.2
